支持在后台全局配置job的过期天数(过期hangfire会自动清理)

+++ Server/Hangfire.HttpJob/HangfireHttpJobOptions.cs
@@ -22,7 +22,7 @@ namespace Hangfire.HttpJob
         /// <summary>
         /// 默认保留7天执行记录
         /// </summary>
-        public int JobExpirationTimeoutDay { get; set; } = 7;
+        public long JobExpirationTimeoutDay { get; set; } = 7;
    
         /// <summary>
         /// 设置默认的执行backgroundjob的queue名称
